refactor: Manager 엔티티 위치 이동 및 관련 코드 정리
