+++ api/src/main/java/com/kernel/app/repository/ManagerAuthRepository.java
@@ -1,6 +1,6 @@
 package com.kernel.app.repository;
 
-import com.kernel.common.admin.entity.Manager;
+import com.kernel.common.manager.entity.Manager;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
