+++ common/src/main/java/com/kernel/common/admin/repository/ManagerRepository.java
@@ -1,6 +1,6 @@
 package com.kernel.common.admin.repository;
 
-import com.kernel.common.admin.entity.Manager;  // 추후에 Manager entity가 정의된 위치로 변경 필요
+import com.kernel.common.manager.entity.Manager;  // 추후에 Manager entity가 정의된 위치로 변경 필요
 import com.kernel.common.admin.entity.Status;   // 추후에 Status entity가 정의된 위치로 변경 필요
 
 import org.springframework.data.jpa.repository.JpaRepository;
